# fitsphere-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitSphere API",
    "description": "Fitness Management Backend",
    "version": "1.0.0"
  },
  "paths": {
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Users",
        "operationId": "get_users_users__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "type": "array",
                  "title": "Response Get Users Users  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "operationId": "create_user_users__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/workouts/{workout_id}/exercises/": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "summary": "Create Exercise",
        "operationId": "create_exercise_workouts__workout_id__exercises__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Get Exercises",
        "operationId": "get_exercises_workouts__workout_id__exercises__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseResponse"
                  },
                  "title": "Response Get Exercises Workouts  Workout Id  Exercises  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workouts/{workout_id}/exercises/{exercise_id}": {
      "delete": {
        "tags": [
          "Exercises"
        ],
        "summary": "Delete Exercise",
        "operationId": "delete_exercise_workouts__workout_id__exercises__exercise_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workout Id"
            }
          },
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/diets/": {
      "get": {
        "tags": [
          "Diet Management"
        ],
        "summary": "Get Diets",
        "operationId": "get_diets_diets__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DietResponse"
                  },
                  "type": "array",
                  "title": "Response Get Diets Diets  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Diet Management"
        ],
        "summary": "Create Diet",
        "operationId": "create_diet_diets__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DietCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DietResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/diets/summary/today": {
      "get": {
        "tags": [
          "Diet Management"
        ],
        "summary": "Today Summary",
        "operationId": "today_summary_diets_summary_today_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/diets/{diet_id}": {
      "get": {
        "tags": [
          "Diet Management"
        ],
        "summary": "Get Diet",
        "operationId": "get_diet_diets__diet_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "diet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DietResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Diet Management"
        ],
        "summary": "Update Diet",
        "operationId": "update_diet_diets__diet_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "diet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DietCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DietResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Diet Management"
        ],
        "summary": "Delete Diet",
        "operationId": "delete_diet_diets__diet_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "diet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/me/profile-picture": {
      "post": {
        "tags": [
          "Profile Picture"
        ],
        "summary": "Upload Profile Picture",
        "operationId": "upload_profile_picture_users_me_profile_picture_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_profile_picture_users_me_profile_picture_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Profile Picture"
        ],
        "summary": "Delete Profile Picture",
        "operationId": "delete_profile_picture_users_me_profile_picture_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_profile_picture_users_me_profile_picture_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_profile_picture_users_me_profile_picture_post"
      },
...
      "DietCreate": {
        "properties": {
          "food_name": {
            "type": "string",
            "title": "Food Name"
          },
          "meal_type": {
            "type": "string",
            "title": "Meal Type"
          },
          "calories": {
            "type": "number",
            "minimum": 0.0,
            "title": "Calories"
          },
          "protein": {
            "type": "number",
            "minimum": 0.0,
            "title": "Protein",
            "default": 0
          },
          "carbohydrates": {
            "type": "number",
            "minimum": 0.0,
            "title": "Carbohydrates",
            "default": 0
          },
          "fats": {
            "type": "number",
            "minimum": 0.0,
            "title": "Fats",
            "default": 0
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "food_name",
          "meal_type",
          "calories",
          "date",
          "user_id"
        ],
        "title": "DietCreate"
      },
      "DietResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "food_name": {
            "type": "string",
            "title": "Food Name"
          },
          "meal_type": {
            "type": "string",
            "title": "Meal Type"
          },
          "calories": {
            "type": "number",
            "title": "Calories"
          },
          "protein": {
            "type": "number",
            "title": "Protein"
          },
          "carbohydrates": {
            "type": "number",
            "title": "Carbohydrates"
          },
          "fats": {
            "type": "number",
            "title": "Fats"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "food_name",
          "meal_type",
          "calories",
          "protein",
          "carbohydrates",
          "fats",
          "date",
          "user_id"
        ],
        "title": "DietResponse"
      },
      "ExerciseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sets": {
            "type": "integer",
            "title": "Sets"
          },
          "reps": {
            "type": "integer",
            "title": "Reps"
          }
        },
        "type": "object",
        "required": [
          "name",
          "sets",
          "reps"
        ],
        "title": "ExerciseCreate"
      },
      "ExerciseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sets": {
            "type": "integer",
            "title": "Sets"
          },
          "reps": {
            "type": "integer",
            "title": "Reps"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sets",
          "reps"
        ],
        "title": "ExerciseResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "age",
          "gender",
          "height",
          "weight",
          "goal",
          "role"
        ],
        "title": "UserResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}